Modern version of gainsight script
