+++ scripts/gainsight/README.md
@@ -1,0 +1,13 @@
+# Original Gainsight script from Gainsight Admin
+
+Below is the original script from gainsight admin, which we make more modern and use in `gainsight.js`
+
+```js
+<script type="text/javascript">
+  (function(n,t,a,e,co){var i="aptrinsic";n[i]=n[i]||function(){
+      (n[i].q=n[i].q||[]).push(arguments)},n[i].p=e;n[i].c=co;
+    var r=t.createElement("script");r.async=!0,r.src=a+"?a="+e;
+    var c=t.getElementsByTagName("script")[0];c.parentNode.insertBefore(r,c)
+  })(window,document,"https://web-sdk.aptrinsic.com/api/aptrinsic.js","AP-PCBATQJJQHRG-2");
+</script>
+```
